Extract contacts extractor logic for re-use
https://trello.com/c/ibAvwvH7

EditionDependenciesExtractor will require the same logic to
scan for contacts within govspeak and populate the dependencies
table.
