+++ lib/govspeak/contacts_extractor.rb
@@ -1,0 +1,19 @@
+module Govspeak
+  EMBEDDED_CONTACT_REGEXP = /\[Contact\:([0-9]+)\]/
+
+  class ContactsExtractor
+    def initialize(govspeak)
+      @govspeak = govspeak
+    end
+
+    def contacts
+      return [] if @govspeak.blank?
+      # scan yields an array of capture groups for each match
+      # so "[Contact:1] is now [Contact:2]" => [["1"], ["2"]]
+      @govspeak.scan(Govspeak::EMBEDDED_CONTACT_REGEXP).map { |capture|
+        contact_id = capture.first
+        Contact.find_by(id: contact_id)
+      }.compact
+    end
+  end
+end
