+++ lib/govspeak/contacts_extractor_helpers.rb
@@ -1,0 +1,7 @@
+module Govspeak
+  module ContactsExtractorHelpers
+    def govspeak_embedded_contacts(govspeak)
+      ContactsExtractor.new(govspeak).contacts
+    end
+  end
+end
